pyproject.toml
[project]
name = "neuro"
version = "0.1.0"
description = "NeuroAvalia backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6.1",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.3.0",
    "django-tailwind>=3.8.0",
    "dj-database-url>=2.2.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.3",
    "pydantic[email]>=2.10.0",
    "python-dateutil>=2.9.0.post0",
    "whitenoise>=6.8.2",
]
